Implement pasting from the clipboard
Implement pasting from the clipboard. As part of this, move the
MacOS key translations, that were being done in KeyDownUp(), to
DoKeyBoardEvent() instead, so we can reuse them more easily in the
pasting code. Move the 'copying to the clipboard' code from coco3.c
to vccclip.cpp, so it lives in the same place as the pasting code.
Fix up the makefiles so C++ is not needed if we're not compiling
the clipboard code.